refactor: move modelPathProps function to cli package
