+++ packages/cli/src/sde-causes.js
@@ -4,7 +4,7 @@ import { parseModel } from '@sdeverywhere/compile'
 
 import { codeGenerator } from './CodeGen.js'
 import { preprocessModel } from './Preprocessor.js'
-import { modelPathProps } from './Helpers.js'
+import { modelPathProps } from './utils.js'
 
 let command = 'causes [options] <model> <C_varname>'
 let describe = 'print dependencies for a C variable name'
